refactor(init): move wizard from services layer to cli/init/wizard
The init wizard (callback, contest, infra, problems) lived in
core/services/init/ but its content is pure CLI/UX: it instantiates
Rich Console, calls ask()/ask_bool() prompts, renders tables, and
prints next-step instructions. None of it is business logic.

Moved verbatim to dom/cli/init/wizard/. The orchestrating function
is renamed callback() -> run_wizard() to read sensibly at its new
location.

Also dropped InitializeProjectOperation. It was a SimpleOperation-
shaped wrapper around init_service_callback() whose only job was to
catch exceptions and produce an OperationResult. Once the wizard
moved to the CLI layer, the operations layer could no longer import
it (would break the layered-architecture contract), and the wrapper
had no remaining purpose. The init command now invokes run_wizard()
directly with inline dry-run and error handling.

Net effect: the services layer no longer depends on Rich; init's
imports correctly flow CLI -> templates/utils.
