+++ dom/cli/init/wizard/__init__.py
@@ -1,0 +1,3 @@
+from .callback import run_wizard
+
+__all__ = ["run_wizard"]
