--- dom/core/operations/init/__init__.py
@@ -1,7 +1,0 @@
-"""Init operations for DomJudge CLI."""
-
-from .initialize_project import InitializeProjectOperation
-
-__all__ = [
-    "InitializeProjectOperation",
-]
