--- dom/core/services/init/__init__.py
@@ -1,1 +1,0 @@
-from .callback import callback
